# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: PolyStack Kit — AI Prompt Task Hub API
  version: 1.0.0
  description: Unified REST API implemented across Django, Laravel, and Flask backends.

servers:
  - url: http://localhost:8000
    description: Django (default dev)
  - url: http://localhost:8080
    description: Flask (example)
  - url: http://localhost
    description: Laravel (XAMPP / artisan serve)

tags:
  - name: Auth
  - name: Tasks
  - name: Executions
  - name: Ratings

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    EnvelopeSuccess:
      type: object
      required: [success, message, data, error]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data: {}
        error:
          nullable: true
    EnvelopeError:
      type: object
      required: [success, message, data, error]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          nullable: true
        error:
          type: object
          properties:
            code:
              type: integer
            details: {}
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
        created_at: { type: string, format: date-time }
    AuthTokens:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
    Task:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        title: { type: string }
        prompt_body: { type: string }
        description: { type: string, nullable: true }
        is_reusable: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TaskInput:
      type: object
      required: [title, prompt_body]
      properties:
        title: { type: string }
        prompt_body: { type: string }
        description: { type: string, nullable: true }
        is_reusable: { type: boolean, default: true }
    Execution:
      type: object
      properties:
        id: { type: integer }
        task_id: { type: integer }
        user_id: { type: integer }
        input_payload: { type: string }
        output_payload: { type: string }
        meta: { nullable: true }
        created_at: { type: string, format: date-time }
    ExecutionInput:
      type: object
      required: [task_id, input_payload, output_payload]
      properties:
        task_id: { type: integer }
        input_payload: { type: string }
        output_payload: { type: string }
        meta: { nullable: true }
    Rating:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        execution_id: { type: integer }
        score: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    RatingInput:
      type: object
      required: [execution_id, score]
      properties:
        execution_id: { type: integer }
        score: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeSuccess'
              examples:
                ok:
                  value:
                    success: true
                    message: Created
                    data:
                      user: { id: 1, email: a@b.com, name: A }
                      tokens: { access_token: eyJ..., token_type: Bearer }
                    error: null
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvelopeError' }

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnvelopeSuccess' }
        '401':
          description: Invalid credentials

  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized

  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
    post:
      tags: [Tasks]
      summary: Create task
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '201':
          description: Created

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '404': { description: Not found }
    put:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '200': { description: OK }
    delete:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: No content (or 200 with envelope per implementation) }

  /api/executions:
    get:
      tags: [Executions]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: task_id
          in: query
          schema: { type: integer }
      responses:
        '200': { description: OK }
    post:
      tags: [Executions]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExecutionInput' }
      responses:
        '201': { description: Created }

  /api/ratings:
    get:
      tags: [Ratings]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: execution_id
          in: query
          schema: { type: integer }
      responses:
        '200': { description: OK }
    post:
      tags: [Ratings]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RatingInput' }
      responses:
        '201': { description: Created }